requirement.txt
## The following requirements were added by pip freeze:
numpy==2.1.0
opencv-python==4.10.0.84
## The following requirements were added by pip freeze:
pillow==10.4.0